requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1